poc: introduce support to multiple cpu nodes
Several steps were needed:
* move the scheduler from MPI_MASTER_THREAD to a new MPI node, MPI_SCHEDULER
* the scheduler now stores a map node:available_gpus
* at the beginning of the program, the scheduler receives from each GPU
  node the number of available GPU
* create a communicator of CPU nodes to allow a graceful termination of
  the program
* add better logging messages

CPU producers ask the scheduler for an available GPU node. If such node exists
it is returned to the sending node and the internal GPU count is updated,
otherwise -1 is returned to communicate the absence of available GPUs,
in which case the CPU producer executes the job by himself.

Upon reception of a valid GPU node the CPU producer promptly send its job to
the GPU node.

The GPU node then updates its internal count of available GPUs and simulates
the job execution. At the end of the computation it notifies the scheduler
of the effective number of free GPUs available on the node.